add github stars banner above chat input
